+++ cincoconfig/encryption.py
@@ -7,7 +7,7 @@
 
 import os
 from itertools import cycle
-from typing import NamedTuple, Optional, Union
+from typing import NamedTuple, Optional, Union, Tuple
 
 try:
     from cryptography.hazmat.primitives.ciphers import Cipher, algorithms, modes
@@ -154,7 +154,7 @@ class KeyFile:
 
         return False
 
-    def _get_provider(self, method: str) -> IEncryptionProvider:
+    def _get_provider(self, method: str) -> Tuple[IEncryptionProvider, str]:
         '''
         Get the encryption provider. ``method`` must be one of
 
@@ -164,15 +164,20 @@ class KeyFile:
           encryption is available (``cryptography`` is installed), :class:`XorProvider` if AES
           is not available
 
-        :returns: the encryption provide instance
+        The resolved method is returned. For example, if ``best`` if specified, the best encryption
+        method will be resolved and returned.
+
+        The return value is a tuple of encryption provider instance and the resolved method.
+
+        :returns: a tuple of ``(provider, method)``
         '''
         if not self.__key:
             raise TypeError('keyfile is not open')
 
         if method == 'aes' or (method == 'best' and AES_AVAILABLE):
-            return AesProvider(self.__key)
+            return AesProvider(self.__key), 'aes'
         if method in ('xor', 'best'):
-            return XorProvider(self.__key)
+            return XorProvider(self.__key), 'xor'
         raise TypeError('invalid encryption method: %s' % method)
 
     def encrypt(self, text: Union[str, bytes], method: str = 'best') -> SecureValue:
@@ -186,7 +191,7 @@ class KeyFile:
 
         bindata = text.encode() if isinstance(text, str) else text
 
-        provider = self._get_provider(method)
+        provider, method = self._get_provider(method)
         ciphertext = provider.encrypt(bindata)
         return SecureValue(method, ciphertext)
 
@@ -198,7 +203,7 @@ class KeyFile:
         if not self.__key:
             raise TypeError('key file is not open')
 
-        provider = self._get_provider(secret.method)
+        provider, _ = self._get_provider(secret.method)
         return provider.decrypt(secret.ciphertext)