+++ tests/test_fields/test_field.py
@@ -4,16 +4,18 @@
 # This file is subject to the terms and conditions defined in the file 'LICENSE', which is part of
 # this source code package.
 #
-
+from unittest.mock import patch
 import pytest
+
 from cincoconfig.abc import Field
 
 
 class MockConfig:
 
-    def __init__(self, data=None):
+    def __init__(self, data=None, parent=None, key=None):
         self._data = data or {}
-
+        self._parent = parent
+        self._key = key
 
 
 class TestBaseField:
@@ -98,5 +100,38 @@ class TestBaseField:
         with pytest.raises(KeyError):
             field.validate(self.cfg, 'hello')
 
+    def test_friendly_name_with_name(self):
+        field = Field(name='ASDF', key='asdf')
+        cfg = MockConfig()
+        assert field.friendly_name(cfg) == 'ASDF'
 
+    def test_friendly_name_same(self):
+        field = Field(name='asdf', key='asdf')
+        cfg = MockConfig()
+        retval = object()
+        with patch.object(field, 'full_path') as mock_path:
+            mock_path.return_value = retval
+            assert field.friendly_name(cfg) is retval
+            mock_path.assert_called_once_with(cfg)
+
+    def test_friendly_name_no_name(self):
+        field = Field(key='asdf')
+        cfg = MockConfig()
+        retval = object()
+        with patch.object(field, 'full_path') as mock_path:
+            mock_path.return_value = retval
+            assert field.friendly_name(cfg) is retval
+            mock_path.assert_called_once_with(cfg)
+
+    def test_full_path_flat(self):
+        field = Field(key='asdf')
+        cfg = MockConfig()
+        assert field.full_path(cfg) == 'asdf'
+
+    def test_full_path_nested(self):
+        root = MockConfig()
+        level1 = MockConfig(key='level1', parent=root)
+        level2 = MockConfig(key='level2', parent=level1)
+        field = Field(key='value')
+        assert field.full_path(level2) == 'level1.level2.value'